refactor: BlockApp, and BlockBase with use cases
